requirements.txt
pytest==2.9.2
pytest-cov==2.3.0
pycrypto==2.6.1